Set up the `laravel-vue-i18n` plugin, replace hardcoded strings with their translations and add route names to all routes

+++ routes/web.php
@@ -22,19 +22,19 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', [GenreController::class, 'index']);
-Route::get('/genre/{genre:slug}', [GenreController::class, 'show']);
+Route::get('/', [GenreController::class, 'index'])->name('genre.index');
+Route::get('/genre/{genre:slug}', [GenreController::class, 'show'])->name('genre.show');
 
-Route::get('/books', [BookController::class, 'index']);
-Route::get('/book/{book:slug}', [BookController::class, 'show']);
+Route::get('/books', [BookController::class, 'index'])->name('book.index');
+Route::get('/book/{book:slug}', [BookController::class, 'show'])->name('book.show');
 
-Route::get('/authors', [AuthorController::class, 'index']);
-Route::get('/author/{author:slug}', [AuthorController::class, 'show']);
+Route::get('/authors', [AuthorController::class, 'index'])->name('author.index');
+Route::get('/author/{author:slug}', [AuthorController::class, 'show'])->name('author.show');
 
-Route::get('/search', [SearchController::class, 'index']);
+Route::get('/search', [SearchController::class, 'index'])->name('search.index');
 
-Route::inertia('/subscribe', 'Subscribe');
-Route::inertia('/terms-and-conditions', 'TermsAndConditions');
+Route::inertia('/subscribe', 'Subscribe')->name('subscribe');
+Route::inertia('/terms-and-conditions', 'TermsAndConditions')->name('terms-and-conditions');
 
 Route::post('/subscription-plan/{id}', [SubscriptionPlanController::class, 'select']);
 
@@ -45,7 +45,7 @@ Route::middleware('auth')->group(function () {
 });
 
 Route::middleware('subscribed')->group(function () {
-	Route::get('/book/{book:slug}/read', [BookController::class, 'read']);
+	Route::get('/book/{book:slug}/read', [BookController::class, 'read'])->name('book.read');
 
 	Route::post('/review/{book_id}', [BookReviewController::class, 'store'])->name('review.store');
 	Route::put('/review/{id}', [BookReviewController::class, 'update'])->name('review.update');
@@ -55,21 +55,21 @@ Route::middleware('subscribed')->group(function () {
 });
 
 Route::middleware('admin')->group(function () {
-	Route::post('/genre', [GenreController::class, 'store']);
-	Route::put('/genre/{id}', [GenreController::class, 'update']);
-	Route::delete('/genre/{id}', [GenreController::class, 'destroy']);
+	Route::post('/genre', [GenreController::class, 'store'])->name('genre.store');
+	Route::put('/genre/{id}', [GenreController::class, 'update'])->name('genre.update');
+	Route::delete('/genre/{id}', [GenreController::class, 'destroy'])->name('genre.destroy');
 
-	Route::post('/book', [BookController::class, 'store']);
-	Route::put('/book/{id}', [BookController::class, 'update']);
-	Route::delete('/book/{id}', [BookController::class, 'destroy']);
+	Route::post('/book', [BookController::class, 'store'])->name('book.store');
+	Route::put('/book/{id}', [BookController::class, 'update'])->name('book.update');
+	Route::delete('/book/{id}', [BookController::class, 'destroy'])->name('book.destroy');
 
-	Route::post('/author', [AuthorController::class, 'store']);
-	Route::put('/author/{id}', [AuthorController::class, 'update']);
-	Route::delete('/author/{id}', [AuthorController::class, 'destroy']);
+	Route::post('/author', [AuthorController::class, 'store'])->name('author.store');
+	Route::put('/author/{id}', [AuthorController::class, 'update'])->name('author.update');
+	Route::delete('/author/{id}', [AuthorController::class, 'destroy'])->name('author.destroy');
 
-	Route::get('/users', [UserController::class, 'index']);
-	Route::put('/user/{id}', [UserController::class, 'update']);
-	Route::delete('/user/{id}', [UserController::class, 'destroy']);
+	Route::get('/users', [UserController::class, 'index'])->name('user.index');
+	Route::put('/user/{id}', [UserController::class, 'update'])->name('user.update');
+	Route::delete('/user/{id}', [UserController::class, 'destroy'])->name('user.destroy');
 });
 
-require __DIR__.'/auth.php';
+require __DIR__ . '/auth.php';
